+++ CHANGELOG.md
@@ -5,6 +5,41 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added dependency tslib.
+- Added dependency @types/react-dom.
+- Added dependency @types/react.
+
+### Changed
+
+- Updated dependency vite-tsconfig-paths to 6.0.4.
+- Updated dependency typescript-eslint to 8.53.0.
+- Updated dependency prettier to 3.8.0.
+- Updated dependency happy-dom to 20.3.0.
+- Updated dependency globals to 17.0.0.
+- Updated dependency eslint-plugin-testing-library to 7.15.4.
+- Updated dependency eslint-plugin-playwright to 2.5.0.
+- Updated dependency eslint-plugin-perfectionist to 5.3.1
+- Updated dependency eslint-plugin-jest to 29.12.1.
+- Updated dependency eslint to 9.39.2.
+- Updated dependency @vitejs/plugin-react to 5.1.2.
+- Updated @types/node to 25.0.8.
+- Updated dependency @testing-library/react to 16.3.1.
+- Updated dependency @eslint/js to 9.39.2.
+- Updated dependency @eslint/compat to 2.0.1.
+- Updated dependency react to 19.2.3.
+- Updated dependency react-dom to 19.2.3.
+- Updated dependency astro to 5.16.9.
+- Converted to Create OutSystems Astro v0.4.0.
+
+### Security
+
+- Updated template dependency undici to 17.8.2 due to CVE-2026-22036.
+- Updated template dependency diff to 8.0.3 due to GHSA-73rr-hh4g-fpgx.
+
 ## [0.1.0] - 2025-12-14
 
 ### Added